+++ scripts/build.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 NAME=tradezero-pricer
-VERSION=0.1.4
+VERSION=0.1.5
 TAG=latest
 PODMAN_ARGS="--layers=false"