+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 CHANGES
 =======
 
+* v 0.1.5 - 09/05/2023
+     * Feat: refactoring api code, add stock factory
+* v 0.1.4 - 01/05/2023
+     * Feat: refactoring all code using blueprints
 * v 0.1.0 - 13/04/2023
      * First commit: Importing repo.